Add a substitute function and other minor changes
The substitute function returns a *new* dLStructure that is the
same as the original, but with RealVariables replaced by Reals
as prescribed by a valuation that is given as an argument.

Also, moved dLFormula.parseFormula to dLStructure.parseStructure,
so all structures can use it. Also changed occurrences of "List"
to "ArrayList". Also made operator and children protected, and
added appropriate getter and setter methods to support this change.